Add mandatory message delivery.
Message will be sent multiple times if no queues are available
to persist. After given timeout an error is raised since no
queues are listening. This confirms the delivery and persistence
of an image to RabbitMQ or raises an error to log.

Test raising error handled by orchestrator. This could
reuse the logging queue to simplify messaging.