Refactor obs service thread usage
sending messages to queues via pika in threads is not
safe. Thus the obs service has been refactored to use
callback methods and handlers. In addition the locking
code has been improved in a way that for lock and
unlock the current state of the metadata is checked
to avoid invalid request errors