Cleanup/Fix logging filter and formatter
Our BaseService class sets up logging format and filter
However that conflicts with the code to use a logfile or
logger through the logging facility. Thus one part of
this patch consolidates logging_logfile with what we
do in the base_service and cleans up the mess

The other part of the patch deals with the log messages.
As we maintain the job_id added by an extra field to
the message record there is no reason why the job_id
needs to be used in the message text which duplicates
that information. Therefore messages sent by the obs
service were adapted

As the same log object is used for logging any kind of
message we should avoid a global filter rule which is able
to silently filter out a message if the log record does
not contain a job_id. Thus this patch also makes sure
no log message can get lost