Added UploadImage class
An instance of UploadImage implements the connection to the
obs service to get the image source as well as the connection
to the credentials and creates the cloud specific Upload object
which actually does the upload. In the still to be done service
class a job will manage one instance of UploadImage in one
thread. The implementation of the credentials handling is still
a noop and might require later refactoring depending on the
way we make use of java web tokens