Added UploadImage class
An instance of UploadImage implements the connection to the
obs and credential services to get the image source and
cloud credentials and creates the cloud specific Upload
object which actually does the upload. In the still to be done
upload service class a job will manage one instance of
UploadImage in one thread.