Refactor use of osc api
After a discussion with Stefan Behlert it's clear that the osc
team has not started to port osc to python3 and it's also clear
that sle15 will provide a py2 interpreter in the legacy module
which can be used by e.g osc during the transition phase. Given
That we want to implement mash for py3 and the osc api is the
only part which is blocking this I have refactored the obs
service to call the osc tool rather than importing the osc
modules as an api.