Fixup setup of self.job_directory
os.makedirs when called with exist_ok returns none if the
directory already exists. Thus self.job_directory is None
in this case but used later methods of the base class
leading to python errors

Also allow consume_queue's queue_name parameter to be
optional. Will be set to self.service_queue if not
specified