No longer use Credentials class in uploader
The original idea was that credentials as they are received
by the credentials service will be put into an instance of
the Credentials class to allow access by methods and encapsulate
that information into a namespace. However the way Sean now
implements the real credentials service prevents this and
so we just pass along whatever data comes from the credentials
service. That way the uploader service holds no reference
to the api provided by the Credentials classes as I wrote
it and hopefully avoids further conflicts. Related to
pull request #88