Delete job description validation from uploader
The validation of the job description has been moved one level
up to the job creator service which applies a json validation
on the request send to the job creator service. It is expected
that the job descriptions created by the job creator service
are valid and doesn't need a validation by the receiving
service. This Fixes #196