Refactor obs service using download server not api
The obs service was based on using the api via the osc
commandline tool. However that requires osc credentials
to be setup which can't be done without opening a wide
gate to access and manipulate the project(s) in obs.
From the mash perspective all we need is access to the
binary build results and its metadata. Both exists on
the published download area. Thus this patch refactors
the obs service to target the download server contents
instead of the accessing the api through osc.
This Fixes #217