Refactor uploader service to be event based
For historic reasons the uploader was based on one consumer
callback method because amqpstorm did not support having
multiple callbacks configured in the consumer. This lead
to an implementation with its own event loop and could be
simplified. In addition some artifacts from former design
aspects were still in the code and got removed by this
refactoring. Related to Issue #337, Fixes #340