Fixup systemd unit files for required services
The startup of the mash services requires the rabbitmq service
to be up and running. Thus it's not enough to run the service
startup after rabbitmq, it must be a required service such that
systemd waits for this service to be active before starting a
mash service. In addition the restart timeout for failed services
is too fast. systemd will stop restarting services if it happens
too fast, therefore a latch of 3sec before the next try is sufficient.
This issue can be seen if you run the mash server in the podman
container from Virtualization:Appliances:Images/mash-image-docker