Cleanup s3 raw image upload.
The cloud_image_name may not always be set. Handle both
cloud_image_name and image_file name. Also, handle different
location paths. Location can be bucket, bucket/path/to/put/file/
and bucket/path/to/file.xz. If a file name is part of the location
string it is used in place of cloud_image_name and image_file.