Update documentation
Added a Quick Start chapter which describes how to use
the mash OCI container. I reworked the OBS project that
builds the mash container to use the latest code base and
to be actually useful again. Project can be found here:
https://build.opensuse.org/project/show/Virtualization:Appliances:Images
This makes using mash and hacking on mash an easier start.